Fixed norm for the builtins and made signals work for heredoc
