+++ includes/builtins.h
@@ -1,5 +1,6 @@
 #ifndef BUILTINS_H
 # define BUILTINS_H
+#include "env.h"
 
 char	*build_path_env(char *input);
 
@@ -31,12 +32,11 @@ int		ft_export(char **av);
 int		ft_unset(char **av);
 int		ft_exit(char **av);
 
-void	spush(t_sstack *stack, t_snode *node);
-t_snode	*spop(t_sstack *stack);
-t_snode	*sinit_node(char *value);
-t_sstack	*sinit_stack(void);
 void	free_str_array(char **str);
 void	update_envp(void);
+void	export_var(char **av);
+char	**populate_strings(t_env *env);
+int		ft_count_nodes(t_env *env);
 
 char	**ft_split(char const *s, char c, int *n);
 int		ft_strcmp(const char *s1, const char *s2);
