+++ includes/signals.h
@@ -10,5 +10,6 @@
 void	activate_signal_handler(void);
 void	sig_handler(int sig);
 void	ignore_signals(void);
+void	here_doc_sig(int sig);
 
 #endif
