+++ srcs/builtins/cd.c
@@ -1,4 +1,16 @@
-# include "minishell.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   cd.c                                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: karpatel <karpatel@student.hive.fi>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/12/17 15:26:50 by karpatel          #+#    #+#             */
+/*   Updated: 2024/12/17 15:26:52 by karpatel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
 
 int	go_home(void)
 {
@@ -23,10 +35,25 @@ int	go_home(void)
 	return (0);
 }
 
+int	update_cd_env(void)
+{
+	char	*cwd;
+
+	add_to_env_list("OLDPWD", get_env("PWD"), 0);
+	cwd = getcwd(NULL, 0);
+	if (!cwd)
+	{
+		printf("minishell: cd: getcwd failed\n");
+		return (0);
+	}
+	add_to_env_list("PWD", gc_strdup(cwd), 1);
+	free(cwd);
+	return (1);
+}
+
 int	ft_cd(char **av)
 {
 	char	*path;
-	char	*cwd;
 
 	if (!av[1])
 		return (go_home());
@@ -38,15 +65,8 @@ int	ft_cd(char **av)
 			printf("bash: cd: %s: No such file or directory\n", av[1]);
 			return (0);
 		}
-		add_to_env_list("OLDPWD", get_env("PWD"), 0);
-		cwd = getcwd(NULL, 0);
-		if (!cwd)
-		{
-			printf("minishell: cd: getcwd failed\n");
+		if (!update_cd_env())
 			return (0);
-		}
-		add_to_env_list("PWD", gc_strdup(cwd), 1);
-		free(cwd);
 	}
 	else
 	{
