+++ srcs/builtins/pwd.c
@@ -1,4 +1,16 @@
-# include "minishell.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   pwd.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: karpatel <karpatel@student.hive.fi>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/12/17 15:27:32 by karpatel          #+#    #+#             */
+/*   Updated: 2024/12/17 15:27:34 by karpatel         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
 
 int	ft_pwd(char **argv)
 {
@@ -6,7 +18,7 @@ int	ft_pwd(char **argv)
 
 	if (argv[0] && argv[1])
 	{
-		printf("minishell: pwd: %s: options and arguments are not allowed\n", argv[1]);
+		printf("minishell: pwd: too many arguments\n");
 		return (1);
 	}
 	pwd = getcwd(NULL, 0);
