--- srcs/parser/fetch_safe.c
@@ -1,20 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   fetch_safe.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: aapadill <aapadill@student.hive.fi>        +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/04 21:38:08 by aapadill          #+#    #+#             */
-/*   Updated: 2024/11/04 21:38:12 by aapadill         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "parsing.h"
-
-int	fetch_top(t_stack *stack)
-{
-	if (stack->top)
-		return (stack->top->value);
-	return (-1);
-}
